requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv

langchain
langchain-openai
langchain-chroma
langchain-text-splitters

chromadb
openai
tiktoken

pydantic